+++ packages/browser/tests/unit/browser.spec.ts
@@ -1,5 +1,5 @@
 import "jest";
-import { Browser } from "../../src";
+import { browser } from "../../src";
 
 import { with_dom, async_with_dom } from "../../../../tests/utils/dom";
 import {
@@ -17,7 +17,7 @@ function run_async_test(test: TestCase) {
       { url: "http://example.com/one" },
       ({ window, resolve }) => {
         test.fn({
-          constructor: Browser,
+          constructor: browser,
           resolve
         });
       }
@@ -29,7 +29,7 @@ function run_test(test: TestCase) {
   it(test.msg, () => {
     with_dom({ url: "http://example.com/one" }, ({ window }) => {
       test.fn({
-        constructor: Browser
+        constructor: browser
       });
     });
   });
@@ -45,10 +45,10 @@ function run_suite(suite: Suite) {
   });
 }
 
-describe("Browser", () => {
+describe("browser", () => {
   it("initializes using window.location", () => {
     with_dom({ url: "http://example.com/one" }, ({ window }) => {
-      const test_history = Browser(pending => {
+      const test_history = browser(pending => {
         pending.finish();
       });
       expect(test_history.location).toMatchObject({
@@ -62,7 +62,7 @@ describe("Browser", () => {
   it("throws if there is no DOM", () => {
     with_dom({ url: "http://example.com/one", set_global: false }, () => {
       expect(() => {
-        const test_history = Browser(pending => {
+        const test_history = browser(pending => {
           pending.finish();
         });
       }).toThrow();
@@ -72,7 +72,7 @@ describe("Browser", () => {
   it('sets initial action to "push" when page has not been previously visited', () => {
     with_dom({ url: "http://example.com/one" }, ({ window }) => {
       window.history.pushState(null, "", "/has-no-key");
-      const test_history = Browser(pending => {
+      const test_history = browser(pending => {
         expect(pending.action).toBe("push");
         pending.finish();
       });
@@ -82,7 +82,7 @@ describe("Browser", () => {
   it('sets initial action to "pop" when page has been previously visited', () => {
     with_dom({ url: "http://example.com/one" }, ({ window }) => {
       window.history.pushState({ key: "17.0" }, "", "/has-key");
-      const test_history = Browser(pending => {
+      const test_history = browser(pending => {
         expect(pending.action).toBe("pop");
         pending.finish();
       });
@@ -102,13 +102,13 @@ describe("go", () => {
   run_suite(go_suite);
 });
 
-describe("Browser history.go", () => {
+describe("browser history.go", () => {
   // integration?
   it("calls window.history.go with provided value", () => {
     with_dom({ url: "http://example.com/one" }, ({ window }) => {
       const real_go = window.history.go;
       const mock_go = (window.history.go = jest.fn());
-      const test_history = Browser(pending => {
+      const test_history = browser(pending => {
         pending.finish();
       });
 
@@ -123,7 +123,7 @@ describe("Browser history.go", () => {
 describe("to_href", () => {
   it("returns the location formatted as a string", () => {
     with_dom({ url: "http://example.com/one" }, () => {
-      const test_history = Browser(pending => {
+      const test_history = browser(pending => {
         pending.finish();
       });
       const path = test_history.to_href({