+++ packages/hash/types/index.d.ts
@@ -1,4 +1,4 @@
 import { ResponseHandler } from "@hickory/root";
 import { HashOptions, HashHistory } from "./types";
 export * from "./types";
-export declare function Hash(fn: ResponseHandler, options?: HashOptions): HashHistory;
+export declare function hash(fn: ResponseHandler, options?: HashOptions): HashHistory;