+++ packages/in-memory/types/in_memory.d.ts
@@ -1,3 +1,3 @@
 import { ResponseHandler } from "@hickory/root";
 import { InMemoryOptions, InMemoryHistory } from "./types";
-export declare function InMemory(fn: ResponseHandler, options?: InMemoryOptions): InMemoryHistory;
+export declare function in_memory(fn: ResponseHandler, options?: InMemoryOptions): InMemoryHistory;